[develop2] refactor around cli and api
