+++ conan/api/conan_api.py
@@ -1,6 +1,7 @@
 import sys
 
 from conan.api.subapi.cache import CacheAPI
+from conan.api.subapi.local import LocalAPI
 from conan.api.subapi.lockfile import LockfileAPI
 from conans import __version__ as client_version
 from conan.api.subapi.config import ConfigAPI
@@ -23,7 +24,7 @@ from conans.model.version import Version
 from conans.paths import get_conan_user_home
 
 
-class ConanAPIV2(object):
+class ConanAPI(object):
     def __init__(self, cache_folder=None):
 
         version = sys.version_info
@@ -56,9 +57,7 @@ class ConanAPIV2(object):
         self.download = DownloadAPI(self)
         self.cache = CacheAPI(self)
         self.lockfile = LockfileAPI(self)
-
-
-ConanAPI = ConanAPIV2
+        self.local = LocalAPI(self)
 
 
 def set_conan_output_level(level, activate_logger=False):
