+++ conan/cli/commands/create.py
@@ -5,7 +5,6 @@ import shutil
 from conan.api.output import ConanOutput, cli_out_write
 from conan.cli.command import conan_command, OnceArgument
 from conan.cli.commands.export import common_args_export
-from conan.cli.commands.install import _get_conanfile_path
 from conan.cli.common import scope_options
 from conan.cli.args import add_lockfile_args, _add_common_install_arguments, _help_build_policies
 from conan.api.conan_app import ConanApp
@@ -37,7 +36,7 @@ def create(conan_api, parser, *args):
     args = parser.parse_args(*args)
 
     cwd = os.getcwd()
-    path = _get_conanfile_path(args.path, cwd, py=True)
+    path = conan_api.local.get_conanfile_path(args.path, cwd, py=True)
     # Now if parameter --test-folder=None (string None) we have to skip tests
     test_folder = False if args.test_folder == "None" else args.test_folder
     test_conanfile_path = _get_test_conanfile_path(test_folder, path)
