+++ conan/cli/commands/install.py
@@ -5,8 +5,7 @@ from conan.api.output import ConanOutput, cli_out_write
 from conan.cli.command import conan_command
 from conan.cli.commands import make_abs_path
 from conan.cli.common import scope_options
-from conan.cli.args import add_lockfile_args, _add_common_install_arguments, add_reference_args, \
-    _help_build_policies
+from conan.cli.args import common_graph_args
 from conan.cli.printers.graph import print_graph_basic, print_graph_packages
 
 from conans.errors import ConanException
@@ -18,43 +17,9 @@ def json_install(info):
     cli_out_write(json.dumps({"graph": deps_graph.serialize()}, indent=4))
 
 
-def _get_conanfile_path(path, cwd, py):
-    """
-    param py= True: Must be .py, False: Must be .txt, None: Try .py, then .txt
-    """
-    candidate_paths = list()
-    path = make_abs_path(path, cwd)
-
-    if os.path.isdir(path):  # Can be a folder
-        if py:
-            path = os.path.join(path, "conanfile.py")
-            candidate_paths.append(path)
-        elif py is False:
-            path = os.path.join(path, "conanfile.txt")
-            candidate_paths.append(path)
-        else:
-            path_py = os.path.join(path, "conanfile.py")
-            candidate_paths.append(path_py)
-            if os.path.exists(path_py):
-                path = path_py
-            else:
-                path = os.path.join(path, "conanfile.txt")
-                candidate_paths.append(path)
-    else:
-        candidate_paths.append(path)
-
-    if not os.path.isfile(path):  # Must exist
-        raise ConanException("Conanfile not found at %s" % " or ".join(candidate_paths))
-
-    if py and not path.endswith(".py"):
-        raise ConanException("A conanfile.py is needed, " + path + " is not acceptable")
-
-    return path
-
-
 def graph_compute(args, conan_api, partial=False, allow_error=False):
     cwd = os.getcwd()
-    path = _get_conanfile_path(args.path, cwd, py=None) if args.path else None
+    path = conan_api.local.get_conanfile_path(args.path, cwd, py=None) if args.path else None
 
     requires = [RecipeReference.loads(r) for r in args.requires] \
         if ("requires" in args and args.requires) else None
@@ -116,20 +81,6 @@ def graph_compute(args, conan_api, partial=False, allow_error=False):
     return deps_graph, lockfile
 
 
-def common_graph_args(subparser):
-    subparser.add_argument("path", nargs="?",
-                           help="Path to a folder containing a recipe (conanfile.py "
-                                "or conanfile.txt) or to a recipe file. e.g., "
-                                "./my_project/conanfile.txt.")
-    add_reference_args(subparser)
-    subparser.add_argument("--requires", action="append",
-                           help='Directly provide requires instead of a conanfile')
-    subparser.add_argument("--tool-requires", action='append',
-                           help='Directly provide tool-requires instead of a conanfile')
-    _add_common_install_arguments(subparser, build_help=_help_build_policies.format("never"))
-    add_lockfile_args(subparser)
-
-
 @conan_command(group="Consumer", formatters={"json": json_install})
 def install(conan_api, parser, *args):
     """
@@ -161,7 +112,7 @@ def install(conan_api, parser, *args):
 
     cwd = os.getcwd()
     if args.path:
-        path = _get_conanfile_path(args.path, cwd, py=None)
+        path = conan_api.local.get_conanfile_path(args.path, cwd, py=None)
         source_folder = os.path.dirname(path)
     else:
         source_folder = cwd
