+++ conan/cli/commands/lock.py
@@ -3,7 +3,8 @@ import os
 from conan.api.output import ConanOutput
 from conan.cli.command import conan_command, OnceArgument, conan_subcommand
 from conan.cli.commands import make_abs_path
-from conan.cli.commands.install import common_graph_args, graph_compute
+from conan.cli.commands.install import graph_compute
+from conan.cli.args import common_graph_args
 from conans.errors import ConanException
 from conans.model.graph_lock import Lockfile, LOCKFILE
 from conans.model.recipe_ref import RecipeReference
