+++ conan/cli/commands/remove.py
@@ -1,11 +1,11 @@
-from conan.api.conan_api import ConanAPIV2
+from conan.api.conan_api import ConanAPI
 from conan.cli.command import conan_command, OnceArgument
 from conan.internal.api.select_pattern import SelectPattern
 from conans.client.userio import UserInput
 
 
 @conan_command(group="Consumer")
-def remove(conan_api: ConanAPIV2, parser, *args):
+def remove(conan_api: ConanAPI, parser, *args):
     """
     Removes recipes or packages from local cache or a remote.
     - If no remote is specified (-r), the removal will be done in the local conan cache.
