+++ conan/cli/commands/search.py
@@ -1,6 +1,6 @@
 from collections import OrderedDict
 
-from conan.api.conan_api import ConanAPIV2
+from conan.api.conan_api import ConanAPI
 from conan.cli.command import conan_command
 from conan.cli.commands.list import print_list_recipes, default_json_formatter
 
@@ -10,7 +10,7 @@ from conans.errors import ConanException
 
 
 @conan_command(group="Consumer", formatters={"text": print_list_recipes, "json": default_json_formatter})
-def search(conan_api: ConanAPIV2, parser, *args):
+def search(conan_api: ConanAPI, parser, *args):
     """
     Searches for package recipes in a remote or remotes
     """
