+++ conans/test/integration/graph/core/graph_manager_base.py
@@ -2,7 +2,7 @@ import os
 import textwrap
 import unittest
 
-from conan.api.conan_api import ConanAPIV2
+from conan.api.conan_api import ConanAPI
 from conans.client.cache.cache import ClientCache
 from conans.model.manifest import FileTreeManifest
 from conans.model.options import Options
@@ -117,7 +117,7 @@ class GraphManagerTest(unittest.TestCase):
         profile_build.process_settings(self.cache)
         build_mode = []  # Means build all
 
-        conan_api = ConanAPIV2(cache_folder=self.cache_folder)
+        conan_api = ConanAPI(cache_folder=self.cache_folder)
 
         root_node = conan_api.graph.load_root_consumer_conanfile(path, profile_host, profile_build)
         deps_graph = conan_api.graph.load_graph(root_node, profile_host, profile_build)
